pyproject.toml
[project]
name = "speakers"
version = "1.0.0"
description = "Speaker diarization app using pyannote and gradio with GPU acceleration"
requires-python = ">=3.11,<3.13"
dependencies = [
    "asteroid-filterbanks>=0.4.0",
    "einops>=0.8.1",
    "huggingface-hub>=0.16.0,<0.19",
    "lightning",
    "pyannote-pipeline",
    "pyannote.audio",
    "pytorch-metric-learning>=2.8.1",
    "rich>=13.9.4",
    "safetensors>=0.5.2",
    "torch-audiomentations>=0.12.0",
    "torch",
    "torchaudio",
    "torchmetrics",
    "soundfile>=0.13.1",
    "matplotlib>=3.10.0",
    "python-dotenv==1.2.1",
    "gradio>=4.0.0,<4.30",
    "pyyaml>=6.0",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
